change setlayout() to take the name of the layout
the setlayout() function used to take a pointer to
the layout in the layouts[] array, whose order might
change and it's generally not very nice.
now setlayout() takes the function pointer of the layout
as an argument, which means it will still work as expected
regardless of any changes to the layouts[] array.