new patch: "sametagrule"
this patch allow new windows to spawn in the same
monitor and tag as other windows based on the specified rules.

two new values are added to the rules: sametagid and parentsametagid.
windows with a given parentsametagid get spawned next to the last
opened window that has the same sametagid. set either one to zero to
disable.